fix intrinsic `ffs` for HIP (#1049)
Native HIP function `__ffs` and `__ffsll` is used with the wrong data
type.

+++ include/alpaka/intrinsic/IntrinsicUniformCudaHipBuiltIn.hpp
@@ -91,7 +91,7 @@ namespace alpaka
                     std::int32_t value)
                 -> std::int32_t
                 {
-                    return __ffs(value);
+                    return __ffs(static_cast<int>(value));
                 }
 
                 //-----------------------------------------------------------------------------
@@ -100,7 +100,7 @@ namespace alpaka
                     std::int64_t value)
                 -> std::int32_t
                 {
-                    return __ffsll(value);
+                    return __ffsll(static_cast<long long>(value));
                 }
             };
         }
